@sgrif 2017-02-10
While we're touching these, can we just refactor it to use `information_schema`, so we can share the code between every adapter except SQLite and Oracle? The query is roughly:

```sql
SELECT column_name FROM information_schema.key_column_usage
    WHERE constraint_name IN (
        SELECT table_constraints.constraint_name FROM information_schema.table_constraints WHERE constraint_type = 'PRIMARY KEY'
    ) AND table_name = $1 AND table_schema = $2
```

(If anyone wanted to refactor all of our schema lookup code to use `information_schema`, you can see examples of what all the queries would need to be [here](https://github.com/diesel-rs/diesel/blob/master/diesel_infer_schema/src/information_schema.rs). The only things that need to differ between databases is the name of the column type column, and what the value of `table_schema` needs to be scoped to when no schema was explicitly given)